fix: address pre-PR review findings
Pre-reviewed by 9 parallel agents across docs, Python, security, persistence, async, conventions/logging/resilience, tests, silent-failure/type-design/api-contract/infra, and issue-resolution. Fixes implemented:

(1) CRITICAL persistence: PostgresCostRecordRepository.save() was missing 'currency' from the params dict, so the first Postgres cost-record write would have raised a KeyError. Fixed by threading the field through.

(2) MAJOR tests: add currency-validator coverage for PeriodSpending and DepartmentSpending so the sibling models of AgentSpending are pinned against the same 'record_count > 0 requires currency' invariant.

(3) MAJOR tests: add Hypothesis @example decorators capturing boundary cases (exactly 2 currencies, 3+, uniform single-record, full 30-record uniform stream) so the property regressions persist in CI per CLAUDE.md.

(4) MAJOR tests: negative tests for _build_session_store / _build_lockout_store / _build_refresh_store on unknown DB handles (unit-level, no Docker required). Complements the positive Postgres integration tests.

(5) MAJOR refactor: extract pure helpers (_validate_time_range, _filter_records, _assert_single_currency, _aggregate, _build_agent_spendings) from tracker.py into synthorg.budget._tracker_helpers so tracker.py stays under the 800-line guideline after the #1446 invariant landed.

Full unit suite green: 21,777 passed, 6 skipped, 0 failed. mypy clean across 3040 source files.

Refs #1446, #1443.
